Get ready for the new season
* Update README
* Archive old race and team files
* Generate new race calendars
* minor changes to scrapers
