+++ config/application.rb
@@ -18,5 +18,6 @@ module WheresMyVdsTeam
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
+    config.autoload_paths << "#{Rails.root}/lib"
   end
 end
